Fix - CodinGame: Fix the simulator and non-cached algorithms
Body:
The simulator was not correctly reporting the losing player which made
it very hard to understand which algorithm was actually working and
which one was not.
After this was fixed, I was able to figure out that the issues with the
algorithms and fix them. They all now function correctly and end up
drawing all the time (if caching is not used).
There are still issues with AlphaBetaPruning with Transposition tables.
Will have to fix that.

Footer: